# contracts/openapi/bundled.yaml
...
openapi: 3.1.0
info:
  title: NPC Simulator Backend API
  version: 2026-04-15
  description: |
    Current backend-facing contract snapshot after the Spring Boot bridge cut.
    Core runtime and review DTOs are sourced from generated JSON Schema so the
    OpenAPI contract stays aligned with the TypeScript runtime models.
servers:
  - url: http://localhost:8080
    description: Local Spring Boot backend
  - url: https://api.example.com
    description: Cloud backend placeholder; replace with the AWS backend URL
paths:
  /api/world:
    get:
      operationId: getWorld
      parameters:
        - in: header
          name: x-world-instance-id
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Current world snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorldSnapshot"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /api/interact:
    post:
      operationId: interact
      parameters:
        - in: header
          name: x-world-instance-id
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractionRequest"
      responses:
        "200":
          description: Interaction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteractionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    InteractionRequest:
      "$schema": "http://json-schema.org/draft-07/schema#"
      "$ref": "#/definitions/InteractionRequestPayload"
      definitions:
        InteractionRequestPayload:
          type: "object"
          properties:
            npcId:
              type: "string"
            targetNpcId:
              type:
                - "string"
                - "null"
            inputMode:
              "$ref": "#/definitions/InputMode"
            text:
              type: "string"
            action:
              anyOf:
                -
                  "$ref": "#/definitions/PlayerAction"
                -
                  type: "null"
            playerId:
              type: "string"
          required:
            - "npcId"
            - "targetNpcId"
            - "inputMode"
            - "text"
            - "action"
            - "playerId"
          additionalProperties: false
        InputMode:
          type: "string"
          enum:
            - "free_text"
            - "action"
            - "combined"
        PlayerAction:
          type: "string"
          enum:
            - "make_case"
            - "expose"
            - "appeal"
            - "ally"
            - "deflect"
            - "stall"
            - "confess"
...
    ErrorResponse:
      "$schema": "http://json-schema.org/draft-07/schema#"
      "$ref": "#/definitions/ErrorResponse"
      definitions:
        ErrorResponse:
          type: "object"
          properties:
            message:
              type: "string"
          required:
            - "message"
          additionalProperties: false
...